-settato il nome alla sessione -eliminate le immagini vecchie -raggruppate le immagine in modo da avere meno richieste al server -cambiati i nomi ai js per permettere a apache di riconoscere l'estensione -spostato nella code il meccanismo di creazione link nella freeplayer -aggiunta campi ordinamento nella freeplayer -aggiunta autentificazione nei file del cron -cambiato il footer togliendo le immagini -reso opzionale il sottotitolo -sistemato tpl della mail formazione -i giocatori estratti per array mantengono l'ordine -sistemata la articolo.inc.php che non ordinava -sistemato css
git-svn-id: http://subversion.assembla.com/svn/fantamanajer@13 37f4913f-0a8a-4499-b109-c3fdd495100e

+++ code/backup.code.php
@@ -6,23 +6,24 @@ $fileSystemObj = new fileSystem();
 $path = 'db';
 $backupName = $path.'/backup_'. date("Y-m-d H:i:s") . '.sql' ;
 $backupObj = new MySQLDump(DBNAME,$backupName,FALSE,FALSE);
-//ESEGUO IL BACKUP SETTIMANALE DEL DB
-if($backupObj->dodump())
+
+if(isset($_GET['user']) && trim($_GET['user']) == 'admin' && isset($_GET['password']) && trim($_GET['password']) == md5('omordotuanuoraoarounautodromo'))
 {
-	$backup[0] = 0;
-	$backup[1] = 'Backup eseguito correttamente';
-	$files = $fileSystemObj->getFileIntoFolder($path);
-	rsort($files);
-	if(count($files) > 8)
+	//ESEGUO IL BACKUP SETTIMANALE DEL DB
+	if($backupObj->dodump())
 	{
-		$lastFile = array_pop($files);
-		unlink($path.'/'.$lastFile);
+		$contenttpl->assign('message','Operazione effettuata correttamente');
+		$files = $fileSystemObj->getFileIntoFolder($path);
+		rsort($files);
+		if(count($files) > 8)
+		{
+			$lastFile = array_pop($files);
+			unlink($path.'/'.$lastFile);
+		}
 	}
+	else
+		$contenttpl->assign('message','Si sono verificati degli errori');
 }
 else
-{
-	$backup[0] = 1;
-	$backup[1] = 'Errore nella creazione del backup';
-}
-$contenttpl->assign('backup',$backup);
+	$contenttpl->assign('message','Non sei autorizzato a eseguire l\'operazione');
 ?>
